feat(substrate): implement SPEC-REFACTOR-v0.2 end-to-end
Apply the SPEC-REFACTOR v0.2 plan (§5 phases 1-8) to realign the substrate
with the mathematical invariants and machine-checkable contracts:

- Ingress (v0.2 §3.4): /ingesta consumes raw text/plain with byte-level
  verification (null-byte scan, strict UTF-8), X-Idempotency-Key enforced
  via unique ingestion_queue key; 415/400/503 fail-loud, never fake 200.
- Schema v0.2: epoch_provenance on geodesic_axes and manifold_nodes;
  lifecycle CHECK restricted to the 4 legal states incl. telemetry_error;
  composite PK (id, epoch_provenance) on axes; legacy migration preserves
  pre-existing rows as seq=1 revisions (RENAME -> recreate -> copy -> drop).
- Kernel: extract pure evaluate_gate_v01() into traianus/core.py
  (no SQLite/FastAPI deps); /consolidar reports the gate under
  dual_key_status.topological_key (PROVISIONAL_INFORMATIONAL_SCORE) and
  preserves the dual-key AND law.
- /mutate epoch-append: inserts a complete new basis tagged
  PROSTHETIC_NSM_V2 (re-padded V1 axes + canonical axis) with no UPDATE;
  V1 remains immutable; basis readers filter by active epoch.
- Determinism/offline (M1/M3): pin dependencies and the all-MiniLM-L6-v2
  revision (1110a243) in pyproject.toml, app.py and bootstrap.py; CI
  cache key and prefetch honor the pinned revision.
- Docs sync: AGENTS.md §4.2 lifecycle states; README honest 384D
  positioning + WP1 north star; Superseding Amendment v0.1 in the ADR
  ledger (prevails over ADR-017/022/023/007).
- Tests/harness: new ingesta() fixture, float32 native fake encoder,
  DDL mirror in db_factory, migration tests, epoch/CHECK tests,
  semantics test; tools/audit_harness.py (C1 guard 45%) and new
  tools/validate_c1_semantics.py (WP0, dual-key law over real data)
  wired into CI.

Full suite: 49 passed. C1 guard: 45% (9/20). WP0: 53% (9/17).

+++ tests/helpers/fake_encoder.py
@@ -31,7 +31,7 @@ class FakeSentenceTransformer:
             digest = hashlib.sha256(text.encode("utf-8")).digest()
             seed = int.from_bytes(digest[:8], "little")
             rng = np.random.default_rng(seed)
-            vec = rng.standard_normal(DIMENSION).astype(np.float64)
+            vec = rng.standard_normal(DIMENSION).astype(np.float32)
             norm = np.linalg.norm(vec)
             out.append(vec / norm if norm > 0 else vec)
         return out[0] if single else np.stack(out)
